Refactor project into package
